# src/UnitConversion/ModelShared/BundleFromOpenApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ModelShared (Service)",
    "description": "Data model containing model dependencies of the clients of the current microservice (WebApp, ServiceTest)",
    "version": "3.2.1"
  },
  "paths": {
...
    "/UnitSystem/{id}": {
      "get": {
        "tags": [
          "UnitSystem"
        ],
        "operationId": "GetUnitSystemById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "UnitSystem"
        ],
        "operationId": "PutUnitSystemById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "UnitSystem"
        ],
        "operationId": "DeleteUnitSystemById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/UnitSystemConversionSet": {
      "get": {
        "tags": [
          "UnitSystemConversionSet"
        ],
        "operationId": "GetAllUnitSystemConversionSetId",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UnitSystemConversionSet"
        ],
        "operationId": "PostUnitSystemConversionSet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Model.UnitSystemConversionSet"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Model.UnitSystemConversionSet"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Model.UnitSystemConversionSet"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/UnitSystemConversionSet/{id}": {
      "get": {
        "tags": [
          "UnitSystemConversionSet"
        ],
        "operationId": "GetUnitSystemConversionSetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OSDC.UnitConversion.Model.UnitSystemConversionSet"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "UnitSystemConversionSet"
        ],
        "operationId": "PutUnitSystemConversionSetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Model.UnitSystemConversionSet"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Model.UnitSystemConversionSet"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Model.UnitSystemConversionSet"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "UnitSystemConversionSet"
        ],
        "operationId": "DeleteUnitSystemConversionSetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "OSDC.DotnetLibraries.General.DataManagement.MetaInfo": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "string",
            "format": "uuid"
          },
          "HttpHostName": {
            "type": "string",
            "nullable": true
          },
          "HttpHostBasePath": {
            "type": "string",
            "nullable": true
          },
          "HttpEndPoint": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem": {
        "type": "object",
        "properties": {
          "Choices": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "ID": {
            "type": "string",
            "format": "uuid"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Description": {
            "type": "string",
            "nullable": true
          },
          "IsDefault": {
            "type": "boolean"
          },
          "IsSI": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "OSDC.UnitConversion.Model.QuantityConversion": {
        "type": "object",
        "properties": {
          "QuantityID": {
            "type": "string",
            "format": "uuid"
          },
          "ValueConversionList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OSDC.UnitConversion.Model.ValueConversion"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OSDC.UnitConversion.Model.UnitSystemConversionSet": {
        "type": "object",
        "properties": {
          "MetaInfo": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OSDC.DotnetLibraries.General.DataManagement.MetaInfo"
              }
            ],
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Description": {
            "type": "string",
            "nullable": true
          },
          "CreationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "LastModificationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "UnitSystemInID": {
            "type": "string",
            "format": "uuid"
          },
          "UnitSystemOutID": {
            "type": "string",
            "format": "uuid"
          },
          "QuantityConversionList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OSDC.UnitConversion.Model.QuantityConversion"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OSDC.UnitConversion.Model.ValueConversion": {
        "type": "object",
        "properties": {
          "DataIn": {
            "type": "number",
            "format": "double"
          },
          "DataOut": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "DataOutString": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}